Netty实战

+++ chapter9/src/test/java/nia/chapter9/AbsIntegerEncoderTest.java
@@ -1,13 +1,13 @@
 package nia.chapter9;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
-
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.channel.embedded.EmbeddedChannel;
 import org.junit.Test;
+
+import static junit.framework.TestCase.assertEquals;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 /**
  * @author mawenlong
@@ -26,7 +26,7 @@ public class AbsIntegerEncoderTest {
     assertTrue(channel.finish());
 
     for (int i = 0; i < 10; i++) {
-      assertEquals(i, channel.readOutbound());
+      assertEquals(i, (int) channel.readOutbound());
     }
     assertNull(channel.readOutbound());
